My changes.

+++ IOHIDManager/IOHIDDevice_.h
@@ -50,6 +50,7 @@
 #pragma mark - includes & imports
 
 #include <AvailabilityMacros.h>
+#include <ConditionalMacros.h>
 
 #include "IOHIDLib_.h"
 
@@ -123,7 +124,7 @@ extern "C" {
 	// Returns:	long - the vendor ID for this device
 	//
 	
-	extern long IOHIDDevice_GetVendorID(IOHIDDeviceRef inIOHIDDeviceRef);
+	extern int IOHIDDevice_GetVendorID(IOHIDDeviceRef inIOHIDDeviceRef);
 	
 	//*************************************************************************
 	//
@@ -149,7 +150,7 @@ extern "C" {
 	// Returns:	long - the product ID for this device
 	//
 	
-	extern long IOHIDDevice_GetProductID(IOHIDDeviceRef inIOHIDDeviceRef);
+	extern int IOHIDDevice_GetProductID(IOHIDDeviceRef inIOHIDDeviceRef);
 	
 	//*************************************************************************
 	//
@@ -227,7 +228,7 @@ extern "C" {
 	// Returns:	long - the location ID for this device
 	//
 	
-	extern long IOHIDDevice_GetLocationID(IOHIDDeviceRef inIOHIDDeviceRef);
+	extern int IOHIDDevice_GetLocationID(IOHIDDeviceRef inIOHIDDeviceRef);
 	
 	//*************************************************************************
 	//
